requirements.txt
python-dotenv==1.0.1
requests==2.32.2
websocket-client==1.8.0
spotipy==2.24.0